backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
sqlalchemy
alembic
apscheduler
passlib[bcrypt]
cryptography
python-jose[cryptography]
python-dotenv
PyPDF2
pdfplumber
python-docx
python-pptx
transformers
sentencepiece
huggingface_hub
torch
accelerate
slowapi
pytest
httpx
